chore(array-utils): rename dir
