+++ array-utils/src/plugins/trolling/CatsPlugin.ts
