+++ array-utils/src/plugins/trolling/HeyInternetPlugin.ts
