+++ array-utils/src/plugins/trolling/WcPlugin.ts
