+++ array-utils/src/plugins/trolling/index.ts
